port build to jbuilder
This moves the JavaScript package into its own opam package called `dispatch-js`, and rename the ocamlfind package from `dispatch.js` to `dispatch-js`.

Also support `js_of_ocaml.3.0` while doing this.  This change allows the core package to build with fewer dependencies, and for your code to explicitly depend on the JavaScript library only if it is needed.

Also test OCaml 4.04 in Travis, as well as add multi-distro testing.

fixes #15 as there are no more depopts on js_of_ocaml