constructor: add a create function
The previous interface wasn't great in that required users to know that
they should partially apply the dispatch function. Now, for both regular
route description and the DSL route description, there's constructor to
the same shared type t. With values of this type, users can now call a
single dispatch function to get a handler for a path.